[ADD] lib/memcopy: use glibc version
the kernel's memcpy and memmove is very inefficient. But the glibc
version is
quite fast, in some cases it is 10 times faster than the kernel version.
So I
introduce some memory copy macros and functions of the glibc to improve
the
kernel version's performance.

The strategy of the memory functions is:
1. Copy bytes until the destination pointer is aligned.
2. Copy words in unrolled loops.  If the source and destination are not
   aligned in the same way, use word memory operations, but shift and
merge
   two read words before writing.
3. Copy the few remaining bytes.

Change-Id: I93e58344177b3b2d9ed19cc9584f22b59c8e828e
Signed-off-by: Miao Xie <miaox*******>